server/app-engine/requirements.txt
Flask==1.1.2
google-cloud-datastore==1.15.3
google-auth<2
requests==2.23.0
gtfs-realtime-bindings==0.0.6
six==1.14.0
pycrypto==2.6.1
ecdsa==0.16.0
google-cloud-ndb==1.11.1
google-cloud-storage==1.43.0